refactor(graph): split workflow logic
