+++ packages/graph/src/workflow/utils/layout/start-axis-alignment.ts
@@ -1,0 +1,244 @@
+/*
+ * Hexabot — Fair Core License (FCL-1.0-ALv2)
+ * Copyright (c) 2026 Hexastack.
+ * Full terms: see LICENSE.md.
+ */
+
+import { getNodesBounds, type Edge } from "@xyflow/react";
+
+import { ENodeType, type GraphNode } from "../../types/workflow-node.types";
+import {
+  END_INDICATOR_ID,
+  START_INDICATOR_ID,
+} from "../graph-builder/id-factory";
+import type { GroupMeta } from "../graph-builder/types";
+import { getWorkflowNodeDimensions } from "../node-metrics.utils";
+
+import {
+  average,
+  getAxisCenter,
+  getBoundsSpreadCenter,
+  getGraphNodeDimensions,
+  isHorizontalDirection,
+  type LayoutContext,
+  translateSpread,
+} from "./geometry";
+import {
+  buildAttachmentMaps,
+  collectAttachmentDescendants,
+  collectNodeIdsWithAttachmentDescendants,
+  mapNodesToGroup,
+} from "./graph-maps";
+
+export const alignAllNodesToStartAxis = (
+  nodes: GraphNode[],
+  edges: Edge[],
+  groups: Map<string, GroupMeta>,
+  ctx: LayoutContext,
+) => {
+  const isVertical = !isHorizontalDirection(ctx);
+  const startNode = nodes.find((node) => node.id === START_INDICATOR_ID);
+
+  if (!startNode) {
+    return nodes;
+  }
+
+  const nodesById = new Map(nodes.map((node) => [node.id, node]));
+  const {
+    childrenByParent: attachmentChildrenByParent,
+    parentByChild: attachmentParentByChild,
+  } = buildAttachmentMaps(edges, nodesById);
+  const nodeToOutermostGroup = mapNodesToGroup(groups, "outermost");
+  // Compute targetAxis from the content nodes (groups and top-level non-indicator
+  // nodes), NOT from Start/Stop positions.  Start/Stop will then be moved to
+  // align with the content axis so they share the same perpendicular position
+  // as the group entry/exit ports.
+  //
+  // For each top-level group, use its bounding-box center.
+  // For each top-level non-indicator, non-attachment node, use its node center.
+  // The targetAxis is the average of all these reference centers.
+  const indicatorIds = new Set([START_INDICATOR_ID, END_INDICATOR_ID]);
+  const processedGroupIds = new Set<string>();
+  const referenceCenters: number[] = [];
+  const collectGroupMemberIds = (groupId: string): Set<string> => {
+    const group = groups.get(groupId);
+
+    if (!group) {
+      return new Set();
+    }
+
+    const memberIds = collectNodeIdsWithAttachmentDescendants(
+      group.memberNodeIds,
+      attachmentChildrenByParent,
+      nodesById,
+    );
+
+    groups.forEach((candidateGroup, candidateGroupId) => {
+      if (
+        candidateGroupId === groupId ||
+        candidateGroup.level <= group.level ||
+        !nodesById.has(candidateGroupId)
+      ) {
+        return;
+      }
+
+      const isNestedGroup = [...candidateGroup.memberNodeIds].some((nodeId) =>
+        group.memberNodeIds.has(nodeId),
+      );
+
+      if (isNestedGroup) {
+        memberIds.add(candidateGroupId);
+      }
+    });
+
+    return memberIds;
+  };
+  const getGroupReferenceCenter = (
+    groupId: string,
+    memberIds = collectGroupMemberIds(groupId),
+  ): number | undefined => {
+    const groupOverlay = nodesById.get(groupId);
+
+    if (groupOverlay) {
+      return getAxisCenter(
+        groupOverlay.position,
+        getGraphNodeDimensions(groupOverlay, ctx),
+        isVertical,
+        "spread",
+      );
+    }
+
+    const memberNodes = [...memberIds]
+      .map((id) => nodesById.get(id))
+      .filter((member): member is GraphNode => Boolean(member));
+
+    if (!memberNodes.length) {
+      return;
+    }
+
+    return getBoundsSpreadCenter(getNodesBounds(memberNodes), isVertical);
+  };
+  // Compute targetAxis from the content reference centers.
+  // Walk all nodes once (skipping attachment children, indicators, and GROUP
+  // overlay nodes) and collect the bounding-box center of each top-level group
+  // and the node center of each standalone non-indicator node.  The targetAxis
+  // is the average of these — Start/Stop will be moved to align with it.
+
+  nodes.forEach((node) => {
+    if (indicatorIds.has(node.id)) {
+      return;
+    }
+
+    // Skip GROUP overlay nodes — they duplicate the bounding-box center of
+    // their member nodes which are already counted via nodeToOutermostGroup.
+    if (node.type === ENodeType.GROUP) {
+      return;
+    }
+
+    if (attachmentParentByChild.has(node.id)) {
+      return;
+    }
+
+    const groupId = nodeToOutermostGroup.get(node.id);
+
+    if (groupId) {
+      if (processedGroupIds.has(groupId)) {
+        return;
+      }
+
+      processedGroupIds.add(groupId);
+      const referenceCenter = getGroupReferenceCenter(groupId);
+
+      if (referenceCenter !== undefined) {
+        referenceCenters.push(referenceCenter);
+      }
+
+      return;
+    }
+
+    const dims = getWorkflowNodeDimensions(node.type, ctx.config);
+
+    referenceCenters.push(
+      getAxisCenter(node.position, dims, isVertical, "spread"),
+    );
+  });
+
+  // Fall back to Start's own center when there are no content nodes.
+  const startDims = getWorkflowNodeDimensions(startNode.type, ctx.config);
+  const targetAxis =
+    referenceCenters.length > 0
+      ? average(referenceCenters)
+      : getAxisCenter(startNode.position, startDims, isVertical, "spread");
+  const deltas = new Map<string, number>();
+  const processedGroups = new Set<string>();
+
+  nodes.forEach((node) => {
+    if (attachmentParentByChild.has(node.id)) {
+      return;
+    }
+
+    // GROUP overlay nodes move with the same delta as their members; skip
+    // them here — they're handled below via the groupId path on member nodes.
+    if (node.type === ENodeType.GROUP) {
+      return;
+    }
+
+    const groupId = nodeToOutermostGroup.get(node.id);
+
+    if (groupId) {
+      if (processedGroups.has(groupId)) {
+        return;
+      }
+
+      processedGroups.add(groupId);
+      const memberIds = collectGroupMemberIds(groupId);
+      const referenceCenter = getGroupReferenceCenter(groupId, memberIds);
+
+      if (referenceCenter === undefined) {
+        return;
+      }
+
+      const delta = targetAxis - referenceCenter;
+
+      if (delta === 0) {
+        return;
+      }
+
+      memberIds.forEach((id) => {
+        deltas.set(id, delta);
+      });
+      // Also shift the GROUP overlay node itself (same delta as its members).
+      deltas.set(groupId, delta);
+
+      return;
+    }
+
+    const dims = getWorkflowNodeDimensions(node.type, ctx.config);
+    const nodeCenter = getAxisCenter(node.position, dims, isVertical, "spread");
+    const delta = targetAxis - nodeCenter;
+
+    if (delta === 0) {
+      return;
+    }
+
+    deltas.set(node.id, delta);
+    collectAttachmentDescendants(
+      node.id,
+      attachmentChildrenByParent,
+      nodesById,
+    ).forEach((childId) => deltas.set(childId, delta));
+  });
+
+  return nodes.map((node) => {
+    const delta = deltas.get(node.id);
+
+    if (delta === undefined || delta === 0) {
+      return node;
+    }
+
+    return {
+      ...node,
+      position: translateSpread(node.position, isVertical, delta),
+    };
+  });
+};
